### 2023-12-02 #6111 @editorialbot
[PRE REVIEW]: An Open-Source Tool for Generating Domain-Specific Accelerators for Resource-Constrained Computing
**Submitting author:** @dtkerns (<a href="http://orcid.org/0009-0006-8622-7629">David Kerns</a>)
**Repository:** https://github.com/dtkerns/d2
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
