### 2023-12-02 #6111 @danielskatz
👋 @dtkerns - thanks for your submission. However, I see a number of issues.

First, your paper is about 4 times longer than recommended for a JOSS paper. Please see the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). If you want to continue this potential JOSS submission, you will need to reduce the content of your .md file.

 In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` after making changes to the .md file or when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

But before making changes to the paper or references, we should be sure that this work is in scope for JOSS.  Due to the relatively small amount of code, I'm unsure if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. Can you explain a bit about the C and Python parts of the code?

Additionally, if we do review this, your README will need work. I don't see a lot of the things that are needed, such as a statement of purpose, community contributor guidelines, how the code is supported, etc.  See https://joss.readthedocs.io/en/latest/review_criteria.html
