+++ paper/paper.bib
@@ -51,9 +51,22 @@ numpages = {27},
 keywords = {FPGAs, performance, field-programmable gate arrays, High-level synthesis, hardware/software codesign, synthesis, power}
 }
 
-@INPROCEEDINGS{limaye21_dosage,  author={Limaye, Ankur and Adegbija, Tosiron},  booktitle={2021 IEEE/ACM International Symposium on Low Power Electronics and Design (ISLPED)},   title={DOSAGE: Generating Domain-Specific Accelerators for Resource-Constrained Computing},   year={2021},  volume={},  number={},  pages={1-6},  doi={10.1109/ISLPED52811.2021.limaye21_dosage}}
+@INPROCEEDINGS{limaye21_dosage,
+  author={Limaye, Ankur and Adegbija, Tosiron},
+  booktitle={2021 IEEE/ACM International Symposium on Low Power Electronics and Design (ISLPED)}, 
+  title={DOSAGE: Generating Domain-Specific Accelerators for Resource-Constrained Computing}, 
+  year={2021},
+  volume={},
+  number={},
+  pages={1-6},
+  doi={10.1109/ISLPED52811.2021.9502501}}
 
-@ARTICLE{limay18_hermit,  author={Limaye, Ankur and Adegbija, Tosiron},  journal={IEEE Internet of Things Journal},   title={HERMIT: A Benchmark Suite for the Internet of Medical Things},   year={2018},  volume={5},  number={5},  pages={4212-4222},  doi={10.1109/JIOT.2018.2849859}}
+@ARTICLE{limay18_hermit,
+  author={Limaye, Ankur and Adegbija, Tosiron},
+  journal={IEEE Internet of Things Journal},
+  title={HERMIT: A Benchmark Suite for the Internet of Medical Things},
+  year={2018},  volume={5},  number={5},  pages={4212-4222},
+  doi={10.1109/JIOT.2018.2849859}}
 
 @article{gansner17_graphviz,
   author    = {Emden R. Gansner and
@@ -106,7 +119,12 @@ keywords = {FPGAs, performance, field-programmable gate arrays, High-level synth
     keywords = "llvm, IR"
 }
 
-@INPROCEEDINGS{guthaus01_mibench,  author={Guthaus, M.R. and Ringenberg, J.S. and Ernst, D. and Austin, T.M. and Mudge, T. and Brown, R.B.},  booktitle={Proceedings of the Fourth Annual IEEE International Workshop on Workload Characterization. WWC-4 (Cat. No.01EX538)},   title={MiBench: A free, commercially representative embedded benchmark suite},   year={2001},  volume={},  number={},  pages={3-14},  doi={10.1109/WWC.2001.990739}}
+@INPROCEEDINGS{guthaus01_mibench,
+  author={Guthaus, M.R. and Ringenberg, J.S. and Ernst, D. and Austin, T.M. and Mudge, T. and Brown, R.B.},
+  booktitle={Proceedings of the Fourth Annual IEEE International Workshop on Workload Characterization. WWC-4 (Cat. No.01EX538)},
+  title={MiBench: A free, commercially representative embedded benchmark suite},
+  year={2001},  volume={},  number={},  pages={3-14},
+  doi={10.1109/WWC.2001.990739}}
 
 @INPROCEEDINGS{karageorgos20_hwsw_bci,
   author={Karageorgos, Ioannis and Sriram, Karthik and Veselý, Ján and Wu, Michael and Powell, Marc and Borton, David and Manohar, Rajit and Bhattacharjee, Abhishek},
