+++ paper/paper.md
@@ -65,7 +65,7 @@ It should be run from the top of the source tree (where one would type `make`) a
 
 ## D2 implementation details
 ### The search for superblocks
-The control flow graph (CFG) is easily generated from the LLVM IR [@ref22_llvmwebsiteintro] files. There are existing tools that do it, but we wrote our own to maintain control over the metadata kept within the CFG. We chose the standard Graphviz [@gansner17_graphviz] package to represent the CFG using their .dot file syntax. While the Graphviz package does support many graph manipulations and operations, superblock identification, as we have defined it, is not something it recognizes or manipulates. Therefore, we wrote our own application (sb.cc) that reads the .dot file and iteratively identifies every SB in the CFG file. This application is the heart of the overall **D2** package and a major contribution of this open-source effort. The `sb` application reads the CFG (.dot) file, within which each BB is represented as a node. Using the C++ Standard Template Library (STL) [@{plauger00_c++stl}], a map is constructed of each node with a vector of connected nodes. Once the graph is parsed into the map/tree, it recursively walks the tree identifying SBs by adding one adjacent node at a time and checking if the "One In, One Out" condition of an SB is satisfied. Once identified, the SB is added to a vector of SBs that is output and then ranked in a later stage.
+The control flow graph (CFG) is easily generated from the LLVM IR [@ref22_llvmwebsiteintro] files. There are existing tools that do it, but we wrote our own to maintain control over the metadata kept within the CFG. We chose the standard Graphviz [@gansner17_graphviz] package to represent the CFG using their .dot file syntax. While the Graphviz package does support many graph manipulations and operations, superblock identification, as we have defined it, is not something it recognizes or manipulates. Therefore, we wrote our own application (sb.cc) that reads the .dot file and iteratively identifies every SB in the CFG file. This application is the heart of the overall **D2** package and a major contribution of this open-source effort. The `sb` application reads the CFG (.dot) file, within which each BB is represented as a node. Using the C++ Standard Template Library (STL) [@plauger00_c++stl], a map is constructed of each node with a vector of connected nodes. Once the graph is parsed into the map/tree, it recursively walks the tree identifying SBs by adding one adjacent node at a time and checking if the "One In, One Out" condition of an SB is satisfied. Once identified, the SB is added to a vector of SBs that is output and then ranked in a later stage.
 
 ### Identification of accelerator candidates
 The key to achieving maximum acceleration is to choose areas of the software that are compute-bound. However, the most expensive compute operation can also be useless to accelerate if it is only executed once. Therefore, identifying and accelerating expensive compute operations that are done many times, typically, in a loop, will yield maximum overall acceleration. Further, if an accelerator can be used by multiple applications, the number of accelerators can be reduced. **D2** makes a concerted effort to identify common and frequently used code segments for acceleration.
