### 2023-12-03 #6111 @dtkerns
On Sun, Dec 3, 2023 at 11:48 AM The Open Journals editorial robot <
***@***.***> wrote:

> Hello @dtkerns <https://github.com/dtkerns>, here are the things you can
> ask me to do:
>
>
> # List all available commands
> @editorialbot commands
>
> # Get a list of all editors's GitHub handles
> @editorialbot list editors
>
> # Check the references of the paper for missing DOIs
> @editorialbot check references
>
> # Perform checks on the repository
> @editorialbot check repository
>
> # Adds a checklist for the reviewer using this command
> @editorialbot generate my checklist
>
> # Set a value for branch
> @editorialbot set joss-paper as branch
>
> # Generates the pdf paper
> @editorialbot generate pdf
>
> # Generates a LaTeX preprint file
> @editorialbot generate preprint
>
> # Get a link to the complete list of reviewers
> @editorialbot list reviewers
>
>
