### 2023-12-04 #6111 @danielskatz
@dtkerns - thanks, I'll look forward to hearing more from you soon.

Also, I see two DOIs that still aren't in the bib, so I'm unsure that "I did address the DOI issues" is fully complete.
