+++ paper/paper.bib
@@ -1,14 +1,5 @@
 %You can copy bibtex citations directly from Google Scholar. Enable through: Settings -> Show links to import ...[Bibtex]; Search for paper title and the bibtex link should appear. Click on it, and copy/past here.
 
-@article{zeigler04discrete,
-  title={Discrete event abstraction: An emerging paradigm for modeling complex adaptive systems},
-  author={Zeigler, Bernard P},
-  booktitle={Perspectives on Adaptation in Natural and Artificial Systems},
-  pages={119--141},
-  year={2004},
-  publisher={Oxford University Press}
-}
-
 @inproceedings{guo04_fpgaAdvantages,
   title={A quantitative analysis of the speedup factors of FPGAs over processors},
   author={Guo, Zhi and Najjar, Walid and Vahid, Frank and Vissers, Kees},
