### 2023-12-05 #6111 @dtkerns
Re: the missing DOIs
for 10.1093/oso/9780195162929.003.001  it looks like the text that
referenced that work was removed, so I removed it from paper.bib
for 10.1109/32.221135, I have an email from the author, who recommended the
bib I cited. (see email below, but I'll follow your advice)

After meeting with my advisor (who's delighted you've engaged in dialogue)
I came up with this explanation:

The tool, D2, enables computer architects and researchers to easily
generate domain-specific accelerators (DSAs) given a set of input workloads
based on the super block (SB) granularity. To the best of our knowledge, no
such automation tool currently exists.
D2 is a shell script “driver” that takes two arguments: the top directory
of the user’s source code to be analyzed and a working directory for D2 to
store its intermediate data.

The driver calls make on the user’s source (for which the Makefile should
have been modified previously, per the documentation, to build the .ll/.llg
"targets") The .llvm files and other metadata is copied to the working
directory. The driver then walks through the various python and C++
data-processing programs to determine a list of rank ordered SBs. The user
is then able to evaluate each SB for appropriateness to implement as a DSA.

The sb.cc program (SB identifier) is the heart of D2 and represents the
greatest contribution of this work. It also represents a significant
research effort not well represented by the line count. It encompassed a
non-trivial search for any already-published Control Flow Graph (CFG) SB
identifiers. A second stage of the effort sought to solve the complex
problem of determining what metadata to carry along to utilize the results
successfully in the context of creating a DSA from the output. Finally, the
sb program is actually called twice from within the driver for the
following two distinct operations:
1) identifying the constraint list from each basic block (BB) in the CFG and
2) identifying the SBs from the CFG.

---
Finally, I will clean up the README... Apologies that I overlooked that
before submission.

Dave
---
Hello,
I'd like to cite your original DOT paper... Google scholar finds it here:
https://labs.onb.ac.at/gitlab/wobweger/h_howto/-/raw/8e8aed73e99899f2f2447098343e83cfd6481bba/0_gen/x_gv/dotguide.pdf
then gives a BibTeX as:

@misc{gansner2006drawing,
  title={Drawing graphs with dot},
  author={Gansner, Emden and Koutsofios, Eleftherios and North, Stephen},
  year={2006},
  publisher={Technical report, AT\&T Research. URL http://www. graphviz.
org/Documentation~…}
}

It would be nice if you hosted it (even hidden) on/in
https://graphviz.org/documentation/ ... maybe you already do ;)

thanks,

Dave
--
Hi,

  Actually that paper is on the Graphviz web site here under User's Guides
in Documentation. As that was not externally published, we usually
recommend the second item under the Graphviz Papers for citations. Or for
just the Dot program, there is the article

@article{DBLP:journals/tse/GansnerKNV93,
  author    = {Emden R. Gansner and
               Eleftherios Koutsofios and
               Stephen C. North and
               Kiem{-}Phong Vo},
  title     = {A Technique for Drawing Directed Graphs},
  journal   = {{IEEE} Trans. Software Eng.},
  volume    = {19},
  number    = {3},
  pages     = {214--230},
  year      = {1993},
  url       = {https://doi.org/10.1109/32.221135},
  doi       = {10.1109/32.221135},
  timestamp = {Wed, 17 May 2017 10:56:38 +0200},
  biburl    = {https://dblp.org/rec/journals/tse/GansnerKNV93.bib},
  bibsource = {dblp computer science bibliography, https://dblp.org}
}

   Emden

On Mon, Dec 4, 2023 at 8:18 AM Daniel S. Katz ***@***.***>
wrote:

> @dtkerns <https://github.com/dtkerns> - thanks, I'll look forward to
> hearing more from you soon.
>
> Also, I see two DOIs that still aren't in the bib, so I'm unsure that "I
> did address the DOI issues" is fully complete.
>
