+++ readme.txt
@@ -1,27 +1,50 @@
+# Statement of need
+D2 enables computer architects and researchers to easily generate domain-specific accelerators (DSAs) given a set of input workloads based on the super block (SB) granularity. To the best of our knowledge, no such automation tool currently exists. D2 is a shell script “driver” that takes two arguments: the top directory of the user’s source code to be analyzed and a working directory for D2 to store its intermediate data.
+
+# Installation Instructions
+
+D2 is intended run on a Unix/Linux system. To build D2, cd to the src directory and type `make`, then add the D2 bin directory to your PATH
+
+# Example usage
+
+ - modify your project Makefiles to generate .ll and .llg files (see below)
+ - run the D2 software
+   $ PATH=$D2BASE/bin:$PATH
+   $ d2 $MYPRJBASE /tmp/myproj
+ - results are found in the file sbselected in the output directory
+ - if you modify the constraints file, you can re-run without remaking the .ll/.llg files with the -m option
+
+# Community guidelines
+
+If you'd like to contrubute to D2, please submit your request on github or contact the authors/maintainers
+
+# Functionality documentation
 
 given: a directory structure of source with makefiles
 
 1) modify the makefile(s), 
- typically add to CFLAGS: -S -emit-llvm
+You'll need to set your compiler (CC in the Makefile) to one capable of generating LLVM, e.g `clang`. It it not required to switch to this compiler for your project, but only for this step of identifying SB
+
+typically add to CFLAGS: -S -emit-llvm
 with a condition ifeq ($(ll),1)
 then anywhere it would make a .o you get a .ll file
 also, second pass with a CFLAGS -g with the conditional
 ifeq ($(llg),1)
 to get a .llg file
 
-2) runnning the d2 script, passing the top of the tree
-  it calls make 2x first with ll=1 then a second time w/ llg=1
-
-it then generates 2 files that contain the list of .ll and .llg files
-the .ll files are the Basic Block nodes each containting the code for each node
+2) run the d2 script, passing the top of the tree it calls make twice, first with ll=1 then a second time w/ llg=1
+it generates two files that contain the list of .ll and .llg files
+The .ll files are the Basic Block nodes each containting the code for each node
 the BB's are named with numeric labels, and each line of the node is unique register name
 
+```
 9:
   %10 = sext i32 %3 to i64
   %11 = getelementptr inbounds i8*, i8** %2, i64 %10
   %12 = sext i32 %1 to i64
   %13 = getelementptr inbounds i8*, i8** %0, i64 %12
   br label %14
+```
 
 The .llg files are similar except that they have debugging information that points each BB to a set of lines in the original source code.
 
@@ -42,18 +65,18 @@ this generates a list of constraints (list of functions) called by each BB
 a constraint is a function that must be realised by an accelerator if the BB that calls it
 is chosen to be accelerated.
 
-the user is then presented with each file of constraints, and they should remove any function calls that they want to allow an accelerator to implement. (eg a pure math function like sqrt is fine, but an io function like
-printf should remain in the constraint file) These constraint files are cached (and available to future "what if" mods)
[…]
