+++ src/Makefile
@@ -1,0 +1,10 @@
+install: all
+	mkdir -p ../bin
+	mv sb ../bin
+
+all: sb
+
+sb: sb.cc
+
+clean:
+	$(RM) ../bin/sb
