### 2023-12-06 5abf164
split readme into .md and .txt
