+++ readme.md
@@ -5,6 +5,10 @@ D2 enables computer architects and researchers to easily generate domain-specifi
 
 D2 is intended run on a Unix/Linux system. To build D2, cd to the src directory and type `make`, then add the D2 bin directory to your PATH
 
+# Dependencies
+D2 requires an LLVM compiler (e.g. clang) to be used for LLVM generation. Additionaly, the clang utility `opt` is also used.
+It is not required to completely switch you application to use clang, it is only used for SB intentification.
+
 # Example usage
 
  - modify your project Makefiles to generate .ll and .llg files (see below)
