### 2023-12-07 #6111 @dtkerns
ok. thanks. Let me talk with my advisor and I'll get back to you as soon as
possible.

On Thu, Dec 7, 2023 at 2:12 PM Daniel S. Katz ***@***.***>
wrote:

> @dtkerns <https://github.com/dtkerns> - thanks for your comments and
> changes so far.
>
> I think, that based on what you've said, this is in scope for JOSS review.
>
> However, there are still two issues to be addressed before that can start,
> one small and one big.
>
> The small one is the DOI - it looks like your bib file contains items for
> both gansner17_graphviz and gansner1993technique which are to the same
> work, but one doesn't have a DOI, which is what editorialbot is catching. I
> think you should just remove the gansner1993technique entry from your
> .bib file, and if needed, make sure you reference gansner17_graphviz
> instead.
>
> The big issue is the paper length. As I wrote previously, this is not a
> JOSS paper, as it's far too long: about 4000 words vs JOSS asking for 1000
> max. The two options you have are:
>
>    1. Reduce the size of the paper, more closely following the example
>    paper
>    <https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography>
>    and the sections in it. Other material that's currently in your paper might
>    be put in the repository and pointed to in the paper.
>    2. Submit to another journal, such as JORS
>    <https://openresearchsoftware.metajnl.com> that has more of a focus on
>    both the paper and the software, unlike JOSS, where the purpose of the
>    paper is mostly to point to and introduce the software and the repository,
>    not to fully explain it.
>
> Please let me know which you want to do.
>
