### 2023-12-07 17e5369
Update paper.bib
remove dupe bib
