+++ paper/paper.bib
@@ -95,17 +95,6 @@ keywords = {FPGAs, performance, field-programmable gate arrays, High-level synth
   url={https://graphviz.org/documentation/}
 }
 
-@article{gansner1993technique,
-  title={A technique for drawing directed graphs},
-  author={Gansner, Emden R and Koutsofios, Eleftherios and North, Stephen C and Vo, K-P},
-  journal={IEEE Transactions on Software Engineering},
-  volume={19},
-  number={3},
-  pages={214--230},
-  year={1993},
-  publisher={IEEE}
-}
-
 @online{ref22_llvmwebsiteabs,
     title = "LLVM Language Reference Manual",
     url  = "https://llvm.org/docs/LangRef.html#abstract",
