### 2023-12-08 #6111 @dtkerns
I spoke with my advisor and he'd like me to proceed w/ JOSS, so I will
start trimming. (I deleted the rogue bib) My advisor would like to review
it before you start your review. I plan to work on it tomorrow. I'll give
you a heads up when we think we're there.
Thanks again!
