+++ paper/full_paper.md
@@ -1,0 +1,206 @@
+---
+title: 'An Open-Source Tool for Generating Domain-Specific Accelerators for Resource-Constrained Computing'
+tags:
+ - Domain-specific accelerators
+ - computer-aided design
+ - resource-constrained systems
+ - energy efficient
+authors:
+ - name: David T Kerns
+   affiliation: 1
+ - name: Tosiron Adegbija
+   affiliation: 1
+affiliations:
+- name: Department of Electrical & Computer Engineering, University of Arizona, Tucson, Arizona, USA
+  index: 1
+date: 01 December 2023
+bibliography: paper.bib
+---
+
+# Summary
+
+Previous work [@limaye21_dosage] introduced the Super Block (SB) as a new granularity for designing Domain Specific Accelerators (DSAs). The SB provides a middle ground to the existing extreme granularities of creating DSAs only from Basic Blocks (BBs) or whole functions. This paper describes an Open Source Software (OSS), called **D2**, that implements the SB approach and will enable researchers and developers in the DSA community to easily experiment with this new granularity towards finding an optimal solution to the DSA definition at hand. Our results show that the SB should be taken seriously as a candidate for creating DSAs. A brief background is presented for readers new to the DSA arena.
+
+# Statement of need
+
+The DSA field currently offers few, if any, open-source automation tools. An important challenge in the design of DSAs is identifying what portions of a set of domain programs should be implemented in hardware (i.e., accelerated) to maximize the performance and energy benefits of the DSA. The SB as a new DSA granularity offers real benefits for efficiently addressing this challenge. However, manually identifying these SBs can be prohibitively time-consuming and error-prone. This ground-up re-write as open source from conception hopes to bring the SB construct to the community at large.
+
+# Motivation for this work
+In a post-Moore\'s Law computing environment where IoT and edge computing continue to expand, resource-constrained high-performance computing systems have become even more important. ASICs (application-specific integrated circuits) continue to set the performance bar, but bring along a high development cost, long lead times, and inability to make corrections or changes once manufactured; not to mention a threshold production run to where it even makes sense economically. GPUs solve the problem by throwing more, but cheaper computing power at a task. This requires a typically large parallelization design effort but in the end, the result is less than optimal from a power usage standpoint. When one of the goals is energy efficiency, GPUs fall short. Additionally, GPUs can be considered to be in the class of a general-purpose accelerator, which typically has a large area footprint and/or an underutilized compute resource that is prohibitive in a resource-constrained environment. Alternatively, a Domain Specific Accelerator (DSA) is an accelerator that is customized for common code blocks shared between multiple workloads. A key takeaway of the DSA is that it typically accelerates a smaller portion of the workload, but at higher utilization rates across workloads, resulting in higher performance of the overall system. While a DSA could be realized in an ASIC, we will assume for this paper that ASICs have already been ruled out due to earlier discussed issues. A DSA on a field-programmable gate array (FPGA) offers a good balance between the often conflicting requirements. The drawback to DSAs, and motivation for this work, is a higher effort in hardware/software co-design since the APIs must be tailored for each DSA. This paper introduces open-source automation tools that address this increased co-design effort.
+
+DSAs, as opposed to general purpose architectures or ASICs, continue to be an attractive alternative because they offer a significant boost to performance without the expensive trade-offs and overhead that creating an ASIC or the size and power requirements of a higher performance general purpose CPU [@guo04_fpgaAdvantages]. However, designing DSAs raises new challenges. The need to analyze different workloads, identify similarities between them, generate accelerators, and then implement the accelerators into the current code base are all challenges that must be overcome and where an opportunity for automation exists. Few, if any, current design automation tools exist for this task.
+
[…]
