+++ paper/paper.md
@@ -18,22 +18,17 @@ bibliography: paper.bib
 ---
 
 # Summary
-
 Previous work [@limaye21_dosage] introduced the Super Block (SB) as a new granularity for designing Domain Specific Accelerators (DSAs). The SB provides a middle ground to the existing extreme granularities of creating DSAs only from Basic Blocks (BBs) or whole functions. This paper describes an Open Source Software (OSS), called **D2**, that implements the SB approach and will enable researchers and developers in the DSA community to easily experiment with this new granularity towards finding an optimal solution to the DSA definition at hand. Our results show that the SB should be taken seriously as a candidate for creating DSAs. A brief background is presented for readers new to the DSA arena.
 
 # Statement of need
-
-The DSA field currently offers few, if any, open-source automation tools. An important challenge in the design of DSAs is identifying what portions of a set of domain programs should be implemented in hardware (i.e., accelerated) to maximize the performance and energy benefits of the DSA. The SB as a new DSA granularity offers real benefits for efficiently addressing this challenge. However, manually identifying these SBs can be prohibitively time-consuming and error-prone. This ground-up re-write as open source from conception hopes to bring the SB construct to the community at large. 
- 
-This paper describes an open-source tool for identifying and generating DSAs given a set of input workloads. We describe how it can be used to identify good candidates for acceleration at the SB granularity in a resource-constrained target, no matter what domain the application lives in. The **D2** tool provides an exhaustive list of SBs that can be simulated to determine the optimal size.
+The DSA field currently offers few, if any, open-source automation tools. An important challenge in the design of DSAs is identifying what portions of a set of domain programs should be implemented in hardware (i.e., accelerated) to maximize the performance and energy benefits of the DSA. The SB as a new DSA granularity offers real benefits for efficiently addressing this challenge. However, manually identifying these SBs can be prohibitively time-consuming and error-prone. This ground-up re-write as open source from conception hopes to bring the SB construct to the community at large. The **D2** tool provides an exhaustive list of SBs that can be simulated to determine the optimal size.
 
 # Key aspects of D2
 This is a follow-on work of [@limaye21_dosage] except re-engineered from the ground up, with special emphasis on making it open source. Additionally, the novel parts are:
-
  - user-controlled constraints
  - normalization of BBs
  - ranking at the BB level and then mapping the BB ranking onto SBs
- - maintaining a link back to the source so that the accelerators can be generated directly from the source code rather than the LLVM intermediate representation (IR) files
+ - maintaining a link back to the source so that the accelerators can be generated directly from the C source code rather than the LLVM intermediate representation (IR) files
 
 ![D2 Flow\label{fig:d2flow}](d2_flow.png)
 
@@ -43,26 +38,25 @@ This is a follow-on work of [@limaye21_dosage] except re-engineered from the gro
 ## Target system model
 The target system model is assumed to be a resource-constrained computing system in which the generated DSA can be integrated. For example, this could be a CPU+FPGA system (used in our case studies), wherein the DSA is implemented on the FPGA, while the unaccelerated portions of the target workloads are run on the CPU. In this scenario, the FPGA should be addressable from the CPU such that the identified code that would normally be run by the CPU can be implemented on the FPGA and offloaded from the CPU without a significant bottleneck. This model is commonly found in resource-constrained computing devices where power constraints are a major concern. For example, human-implanted medical devices must be concerned with power and heat dissipation so that the surrounding tissues are not damaged [@karageorgos20_hwsw_bci].
 
-## D2 workflow
[…]
