+++ paper/paper.bib
@@ -1,5 +1,21 @@
 %You can copy bibtex citations directly from Google Scholar. Enable through: Settings -> Show links to import ...[Bibtex]; Search for paper title and the bibtex link should appear. Click on it, and copy/past here.
 
+@book{Hennessy_Patterson_2019,
+  address={New York},
+  title={A new golden age for computer architecture},
+  volume={62},
+  ISSN={0001-0782},
+  DOI={10.1145/3282307},
+  number={2},
+  journal={Communications of the ACM},
+  publisher={ACM},
+  author={Hennessy, John and Patterson, David},
+  year={2019},
+  pages={48–60},
+  keywords={Computer architecture ; Computer interfaces ; Computer security ; Design and construction ; Forecasts and trends ; Innovations ; Interfaces ; Open systems (Computers) ; Software ; Systems design},
+  language={eng}
+}
+
 @inproceedings{guo04_fpgaAdvantages,
   title={A quantitative analysis of the speedup factors of FPGAs over processors},
   author={Guo, Zhi and Najjar, Walid and Vahid, Frank and Vissers, Kees},
@@ -18,11 +34,11 @@
   series = {FPGA '04}
 }
 
-@misc{unspecified21_vivado_hls,
+@misc{xilinx21_vivado_hls,
 title = {Introduction to FPGA Design with Vivado High-Level Synthesis},
 url = {https://docs.xilinx.com/v/u/en-US/ug998-vivado-intro-fpga-design-hls},
 publisher = {Xilinx},
-author = {unspecified},
+author = {Xilinx, I},
 year = {2021},
 abstract = {Software is the basis of all applications. Whether for entertainment, gaming,
 communications, or medicine, many of the products people use today began as a software
