+++ readme.txt
@@ -5,8 +5,8 @@ You'll need to set your compiler (CC in the Makefile) to one capable of generati
 
 typically add to CFLAGS: -S -emit-llvm
 with a condition ifeq ($(ll),1)
-then anywhere it would make a .o you get a .ll file
-also, second pass with a CFLAGS -g with the conditional
+then anywhere it would make a .o it will build a .ll file instead
+also, a second pass with a CFLAGS -g with the conditional
 ifeq ($(llg),1)
 to get a .llg file
 
@@ -33,7 +33,7 @@ opt -analyze --loops
 this custom parser retains information from the .ll file, namely, the function name and label number
 additionally, it keeps track of any functions a BB calls
 and finally what other nodes (BB) the curent BB can branch to
-this form the directed graph or .dot file which can be directly viewed with the dot utility
+this forms the directed graph or .dot file which can be directly viewed with the dot utility
 Finally, for each function a fake entry and exit node are created to serve as "handles" when the function is called from another function else where in the program
 
 5.1) we call the sb program with the -c option on each .dot file
@@ -47,22 +47,25 @@ The user is then presented with each file of constraints, and they should remove
 
 5.3) The single output file is split into one .dot file per SB
 
-6) create a BB database (DB) DB here is a delimited (usually colon) list of fields
+6) create a BB database (DB) DB here is a flat file, delimited (usually colon) list of fields
 ex:
 
 ./lame/lame3.70/quantize.ll:amp_scalefac_bands:58:L_amp_scalefac_bands_270:mul 4:fmul 4:udiv 0:sdiv 0:fdiv 0:urem 0:srem 0:frem 0:
 ./lame/lame3.70/vbrquantize.ll:VBR_iteration_loop_new:26:L_VBR_iteration_loop_new_116:mul 0:fmul 3:udiv 0:sdiv 0:fdiv 0:urem 0:srem 0:frem 0:llvm_fabs_v2f64,llvm_sqrt_v2f64,llvm_sqrt_v2f64,llvm_fabs_v2f64,llvm_sqrt_v2f64,llvm_sqrt_v2f64,llvm_fabs_v2f64,llvm_sqrt_v2f64,llvm_sqrt_v2f64
 
 7) create a loop DB
+ex:
 
 ./jpeg/jpeg-6a/jccoefct.ll:compress_first_pass:L_compress_first_pass_186:3
 
 
 8) create a BB SB DB
+ex:
 
 ./jpeg/jpeg-6a/jcapimin.ll:L_jpeg_finish_compress_10:0005:0007:0008
 
 9) rank order BB (DB)
+ex:
 
 ./jpeg/jpeg-6a/jdcoefct.ll:L_decompress_onepass_108:5:0:0027,0028,0034
 ./jpeg/jpeg-6a/jdcoefct.ll:L_consume_data_98:5:0:0005,0007
@@ -73,6 +76,7 @@ ex:
 11) create a normalized version of the file. in this step we renumber the registers of the the BB to always start at %1 ... this allows us to compare one BB to another to find similar BBs
 
 12) rank order normalized files (DB)
+ex:
 
 ./lame/lame3.70/newmdct.bb/L_mdct_sub48_408.norm
 ./lame/lame3.70/newmdct.bb/L_mdct_sub48_520.norm
@@ -80,9 +84,11 @@ ex:
 ./lame/lame3.70/newmdct.bb/L_mdct_sub48_488.norm
 
 13) identify SB w/ high ranked BBs (DB)
+ex:
 
 ./jpeg/jpeg-6a/jdcoefct.db.d/g0061.dot
[…]
