+++ paper/paper.bib
@@ -178,7 +178,7 @@ numpages = {7}
     keywords = "meld, diff, merge, python"
 }
 
-@book{plauger00_c++stl,
+@book{plauger00_cpp_stl,
 author = {Plauger, P.J. and Lee, Meng and Musser, David and Stepanov, Alexander A.},
 title = {C++ Standard Template Library},
 year = {2000},
