+++ paper/paper.md
@@ -24,7 +24,7 @@ Domain-specific accelerators (DSAs) [@Hennessy_Patterson_2019] are crucial in mo
 The DSA field currently offers few, if any, open-source automation tools. An important challenge in the design of DSAs is identifying what portions of a set of domain programs should be implemented in hardware (i.e., accelerated) to maximize the performance and energy benefits of the DSA. [@Hennessy_Patterson_2019] even argue for a new domain-specific language. The SB as a new DSA granularity offers real benefits for efficiently addressing this challenge. However, manually identifying these SBs can be prohibitively time-consuming and error-prone. This ground-up re-write as open source from conception hopes to bring the SB construct to the community at large. The **D2** tool provides an exhaustive list of SBs that can be simulated to determine the optimal size.
 
 # Key aspects of D2
-This is a follow-on work of [@limaye21_dosage] except re-engineered from the ground up, with special emphasis on making it open source. Additionally, the novel parts are:
+This is a follow-on work of [@limaye21_dosage], re-engineered from the ground up, with special emphasis on making it open source. Additionally, the novel parts are:
 
  - user-controlled constraints
  - normalization of BBs
@@ -34,7 +34,7 @@ This is a follow-on work of [@limaye21_dosage] except re-engineered from the gro
 ![D2 Flow\label{fig:d2flow}](d2_flow.png)
 
 # D2 Overview
-**D2**, as depicted in \autoref{fig:d2flow}, is a tool that accepts a system\'s tree of source code. The user makes minor modifications to the makefiles to generate the required LLVM IR files. **D2** then evaluates the IR files to create a set of metadata files. It then identifies and ranks BBs and SBs as candidates for FPGA acceleration to produce an optimal set of accelerators for the system. The identified source lines are then given to a synthesis tool, such as Xilinx Vivato HLS [@unspecified21_vivado_hls] to produce an FPGA accelerator that is integrated back into the source code. The resulting system runs faster and more efficiently, potentially making the system viable on a resource-constrained target, which without the DSA would not be achievable.
+**D2**, as depicted in \autoref{fig:d2flow}, is a tool that accepts a system\'s tree of source code. The user makes minor modifications to the makefiles to generate the required LLVM IR files. **D2** then evaluates the IR files to create a set of metadata files. It then identifies and ranks BBs and SBs as candidates for FPGA acceleration to produce an optimal set of accelerators for the system. The identified source lines are then given to a synthesis tool, such as Xilinx Vivato HLS [@unspecified21_vivado_hls] to produce an FPGA accelerator that is integrated back into the source code. The resulting system runs faster and more efficiently, potentially making the system even viable on a resource-constrained target.
 
 ## Target system model
 The target is assumed to be a resource-constrained computing system in which the generated DSA can be integrated. For example, this could be a CPU+FPGA system, wherein the DSA is implemented on the FPGA, while the unaccelerated portions of the target workloads are run on the CPU. In this scenario, the FPGA should be addressable from the CPU such that the identified code that would normally be run by the CPU can be implemented on the FPGA and offloaded from the CPU without a significant bottleneck. This model is commonly found in resource-constrained computing devices where power constraints are a major concern, like human-implanted medical devices [@karageorgos20_hwsw_bci], for example.
@@ -43,7 +43,7 @@ The target is assumed to be a resource-constrained computing system in which the
 **D2** is a shell script that calls the individual programs written in both Python and C++ that process the LLVM files generated by the compiler.
 
 ### The search for superblocks
[…]
