### 2023-12-12 #6111 @dtkerns
On Fri, Dec 8, 2023 at 10:53 PM The Open Journals editorial robot <
***@***.***> wrote:

> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1145/968280.968304 is OK
> - 10.1145/2514740 is OK
> - 10.1109/ISLPED52811.2021.9502501 is OK
> - 10.1109/JIOT.2018.2849859 is OK
> - 10.1109/32.221135 is OK
> - 10.1109/WWC.2001.990739 is OK
> - 10.1109/ISCA45697.2020.00041 is OK
> - 10.1145/2024716.2024718 is OK
> - 10.1109/HPCA.2017.59 is OK
>
> MISSING DOIs
>
> - None
>
> INVALID DOIs
>
> - None
>
