### 2023-12-18 #6111 @dtkerns
I added an example directory. When I ran d2 on the example, I discovered
I've been hit by software rot. (the usage of LLVM's opt command changed)
I'm working on correcting the issue.

On Tue, Dec 12, 2023 at 7:15 AM The Open Journals editorial robot <
***@***.***> wrote:

> Assigned! @gkthiruvathukal <https://github.com/gkthiruvathukal> is now
> the editor
>
