### 2024-01-06 cbe719f
order of operations
