### 2024-01-08 82a23e3
fix logic for $(if ) syntax
