### 2024-01-08 612d571
add -d flag for debug output
