### 2024-01-08 #6111 @dtkerns
It turns out to be a bug in LLVM ... it took a significant amount of
patience to discover this. see
https://discourse.llvm.org/t/need-usage-help-w-new-pass-manager-for-opt-analysis-natural-loop-information/75874
which ultimately resulted in
https://github.com/llvm/llvm-project/issues/76762
I've created an issue against d2 and implemented a work-around in the
script.
