### 2024-02-09 #6111 @kinow
> @kinow, and @xirdneh, would either or both of you be able to help with this JOSS submission?

HI @gkthiruvathukal I have a co-worker doing his phd in an area that seems connected to the subject of this paper. I asked him to have a look if he'd like to review this one instead.
