### 2024-02-13 #6111 @kinow
> @kinow Can you make sure your co-worker is on GitHub? I'd like a link to his GitHub profile and his web site, if possible.

Sure, @gkthiruvathukal, @manuel-g-castro, although I think most of the work he did during his masters and recent work from his phd are on this GitLab server: https://earth.bsc.es/gitlab/mgimenez
