### 2024-02-25 #6111 @gkthiruvathukal
@andrewjpage Are you willing to be a second reviewer on this JOSS submission?
