### 2024-03-02 dtkerns/d2#4 @manuel-g-castro
Minor Comments for the JOSS Review
Hello, @dtkerns!

I am currently reviewing your submission to JOSS, in issue openjournals/joss-reviews#6111.  For that, I am using an `ubuntu:latest` docker image.

So far, I have found these small issues:

- I find odd the organization of the information: we have two readme files which differ only by their extension. One contains the usual information: the JOSS required statement of need, installation, dependencies, example, and how to contribute; the other contains information on how to utilize the software. I would suggest to merge them into a single e readme or rename the `readme.txt` file to something else.
- the dependencies section of d2 is rather confusing: at first glance I thought that `opt` was an optional dependency. I would suggest to make it clearer, like "d2 requires an LLVM compiler and opt".
- have you tested with any other LLVM compiler? Or just clang? If the latter is the case, I would also explicitly mention that you have only tested with clang.
- a couple of little typos in `./readme.md`, line 26, it says "contrubute" instead of contribute. In line 9, "Additionaly" instead of additionally.

I will continue to comment here as I progress in the review.
