### 2024-03-02 dtkerns/d2#5 @manuel-g-castro
Replicating Example for JOSS Submission
Hello, @dtkerns:

For the review of your submission to JOSS, openjournals/joss-reviews/issues/6111, I am trying to replicate the examples.

I have pulled the `ubuntu:latest` docker image to run the testing of your software. I had to change the default symlink `cc` to point to `/usr/bin/clang` (in this image it was pointing to `gcc`), otherwise both apps would not be compiled.

But, even with the llvm capable compiler set, running `d2` prints some errors due to files not existing (see log below).  Moreover, all the files produced by it are empty or really tiny.

I am unsure that this is the desired outcome.

```
root@6ce3f6fc9c55:/home/root/d2/output# find -type f -exec ls -l {} \;
-rw-r--r-- 1 root root 0 Mar  2 23:20 ./app1/app1.sb
-rw-r--r-- 1 root root 12 Mar  2 23:20 ./app1/app1.dot
-rw-r--r-- 1 root root 0 Mar  2 23:20 ./app1/app1.loop
-rw-r--r-- 1 root root 0 Mar  2 22:39 ./app1/app1.const
-rw-r--r-- 1 root root 0 Mar  2 23:20 ./bbselected
-rw-r--r-- 1 root root 0 Mar  2 23:20 ./bbsel_norm
-rw-r--r-- 1 root root 0 Mar  2 23:20 ./lobb
-rw-r--r-- 1 root root 0 Mar  2 23:20 ./bbdb
-rw-r--r-- 1 root root 0 Mar  2 23:20 ./bbsel_norm.bb
-rw-r--r-- 1 root root 0 Mar  2 23:20 ./loopdb
-rw-r--r-- 1 root root 30 Mar  2 23:20 ./loll
-rw-r--r-- 1 root root 0 Mar  2 23:20 ./bbsbdb
-rw-r--r-- 1 root root 0 Mar  2 23:20 ./app2/app2.loop
-rw-r--r-- 1 root root 12 Mar  2 23:20 ./app2/app2.dot
-rw-r--r-- 1 root root 0 Mar  2 23:20 ./app2/app2.sb
-rw-r--r-- 1 root root 0 Mar  2 22:33 ./app2/app2.const
-rw-r--r-- 1 root root 0 Mar  2 23:20 ./bbcdb
-rw-r--r-- 1 root root 32 Mar  2 23:20 ./lollg
-rw-r--r-- 1 root root 0 Mar  2 23:20 ./sbselected
-rw-r--r-- 1 root root 0 Mar  2 23:20 ./rankorder
-rw-r--r-- 1 root root 0 Mar  2 23:20 ./ronormfiles
```

```
root@6ce3f6fc9c55:/home/root/d2/example# d2 /home/root/d2/output/
/home/root/d2/bin/d2: line 14: [: 14.0.0: integer expression expected
/home/root/d2/bin
/home/root/d2/output
for d in app1 app2; do \
	make -C $d clean; \
done
make[1]: Entering directory '/home/root/d2/example/app1'
rm -f app1 *.ll *.llg
make[1]: Leaving directory '/home/root/d2/example/app1'
make[1]: Entering directory '/home/root/d2/example/app2'
rm -f app2 *.ll *.llg
make[1]: Leaving directory '/home/root/d2/example/app2'

real	0m0.037s
user	0m0.021s
sys	0m0.015s
make -C app1
make[1]: Entering directory '/home/root/d2/example/app1'
clang -std=c99 -S -emit-llvm app1.c -o app1.ll
make[1]: Leaving directory '/home/root/d2/example/app1'
make -C app2
make[1]: Entering directory '/home/root/d2/example/app2'
clang -std=c99 -S -emit-llvm app2.c -o app2.ll
make[1]: Leaving directory '/home/root/d2/example/app2'

real	0m0.209s
user	0m0.123s
sys	0m0.084s
make -C app1
make[1]: Entering directory '/home/root/d2/example/app1'
clang -std=c99 -S -emit-llvm -S -emit-llvm -g app1.c -o app1.llg
make[1]: Leaving directory '/home/root/d2/example/app1'
make -C app2
make[1]: Entering directory '/home/root/d2/example/app2'
clang -std=c99 -S -emit-llvm -S -emit-llvm -g app2.c -o app2.llg
make[1]: Leaving directory '/home/root/d2/example/app2'

real	0m0.122s
user	0m0.083s
sys	0m0.038s
copy files to /home/root/d2/output
/home/root/d2/bin/d2: line 64: cpio: command not found
create ll2c db
llg2cdb ./app1/app1.llg
Traceback (most recent call last):
  File "/home/root/d2/bin/llg2cdb.py", line 77, in <module>
    llg2cdb(f)
  File "/home/root/d2/bin/llg2cdb.py", line 9, in llg2cdb
    with open(filename, "r") as f:
FileNotFoundError: [Errno 2] No such file or directory: './app1/app1.llg'
llg2cdb ./app2/app2.llg
Traceback (most recent call last):
  File "/home/root/d2/bin/llg2cdb.py", line 77, in <module>
    llg2cdb(f)
  File "/home/root/d2/bin/llg2cdb.py", line 9, in llg2cdb
    with open(filename, "r") as f:
FileNotFoundError: [Errno 2] No such file or directory: './app2/app2.llg'

real	0m0.069s
user	0m0.062s
sys	0m0.009s
create loop files
opt ./app1/app1.ll -passes='print<loops>' -disable-output > ./app1/app1.loop
opt: ./app1/app1.ll: error: Could not open input file: No such file or directory
opt ./app2/app2.ll -passes='print<loops>' -disable-output > ./app2/app2.loop
opt: ./app2/app2.ll: error: Could not open input file: No such file or directory

real	0m0.049s
user	0m0.024s
sys	0m0.024s
create dot files
ll2dot ./app1/app1.ll > ./app1/app1.dot
Traceback (most recent call last):
  File "/home/root/d2/bin/ll2dot.py", line 116, in <module>
    ll2dot(f)
  File "/home/root/d2/bin/ll2dot.py", line 29, in ll2dot
    with open(filename, "r") as f:
FileNotFoundError: [Errno 2] No such file or directory: './app1/app1.ll'
ll2dot ./app2/app2.ll > ./app2/app2.dot
Traceback (most recent call last):
  File "/home/root/d2/bin/ll2dot.py", line 116, in <module>
    ll2dot(f)
  File "/home/root/d2/bin/ll2dot.py", line 29, in ll2dot
    with open(filename, "r") as f:
FileNotFoundError: [Errno 2] No such file or directory: './app2/app2.ll'

real	0m0.075s
user	0m0.057s
sys	0m0.018s
create sb files
/home/root/d2/output
sb -f ./app1/app1.const ./app1/app1.dot > ./app1/app1.sb
sb -f ./app2/app2.const ./app2/app2.dot > ./app2/app2.sb

real	0m0.057s
user	0m0.038s
sys	0m0.018s
create bbdb
/home/root/d2/output
ll2db ./app1/app1.ll
Traceback (most recent call last):
  File "/home/root/d2/bin/ll2db.py", line 143, in <module>
    ll2db(f)
  File "/home/root/d2/bin/ll2db.py", line 31, in ll2db
    with open(filename, "r") as f:
FileNotFoundError: [Errno 2] No such file or directory: './app1/app1.ll'
ll2db ./app2/app2.ll
Traceback (most recent call last):
  File "/home/root/d2/bin/ll2db.py", line 143, in <module>
    ll2db(f)
  File "/home/root/d2/bin/ll2db.py", line 31, in ll2db
    with open(filename, "r") as f:
FileNotFoundError: [Errno 2] No such file or directory: './app2/app2.ll'

real	0m0.084s
user	0m0.072s
sys	0m0.010s
create loopdb
loop2db ./app1/app1.loop
loop2db ./app2/app2.loop

real	0m0.048s
user	0m0.029s
sys	0m0.016s
create bbsbdb
sb2db ./app1/app1.sb
sb2db ./app2/app2.sb

real	0m0.047s
user	0m0.042s
sys	0m0.005s
-rw-r--r-- 1 root root 0 Mar  2 23:20 /home/root/d2/output/bbcdb
-rw-r--r-- 1 root root 0 Mar  2 23:20 /home/root/d2/output/bbdb
-rw-r--r-- 1 root root 0 Mar  2 23:20 /home/root/d2/output/bbsbdb
-rw-r--r-- 1 root root 0 Mar  2 23:20 /home/root/d2/output/loopdb
number bb 0
number bb in loops 0
rank SB

real	0m0.024s
user	0m0.020s
sys	0m0.005s
separate ll into bb files
Traceback (most recent call last):
  File "/home/root/d2/bin/ll2bb.py", line 54, in <module>
    ll2bb(f)
  File "/home/root/d2/bin/ll2bb.py", line 9, in ll2bb
    with open(file, "r") as f:
FileNotFoundError: [Errno 2] No such file or directory: './app1/app1.ll'

real	0m0.039s
user	0m0.037s
sys	0m0.001s
create normalized file for each bb file

real	0m0.001s
user	0m0.001s
sys	0m0.000s
rank order normalized files

real	0m0.001s
user	0m0.001s
sys	0m0.000s
bbselect

real	0m0.044s
user	0m0.033s
sys	0m0.010s
ID SB from BB

real	0m0.001s
user	0m0.000s
sys	0m0.003s
colorize BB within SB CFG

real	0m0.001s
user	0m0.001s
sys	0m0.000s
```
