+++ step-by-step.md
@@ -1,95 +1,108 @@
+# Step-by-step
+
 given: a directory structure of source with makefiles
 
 1) modify the makefile(s), 
-You'll need to set your compiler (CC in the Makefile) to one capable of generating LLVM, e.g `clang`. It it not required to switch to this compiler for your project, but only for this step of identifying SB
+You'll need to set your compiler (`CC` in the Makefile) to one capable of generating LLVM, e.g `clang`. It it not required to switch to this compiler for your project, but only for this step of identifying SB
 
-typically add to CFLAGS: -S -emit-llvm
-with a condition ifeq ($(ll),1)
+typically add to CFLAGS: `-S -emit-llvm`
+with a condition `ifeq ($(ll),1)`
 then anywhere it would make a .o it will build a .ll file instead
-also, a second pass with a CFLAGS -g with the conditional
-ifeq ($(llg),1)
+also, a second pass with a CFLAGS `-g` with the conditional
+`ifeq ($(llg),1)`
 to get a .llg file
 
-2) run the d2 script from the top of the tree with the path to the output directory as an argument.
-it calls make twice, first with ll=1 then a second time w/ llg=1
-it generates two files that contain the list of .ll and .llg files
-The .ll files are the Basic Block nodes each containting the code for each node
-the BB's are named with numeric labels, and each line of the node is unique register name
+2) Run the `d2` script from the top of the tree with the path to the output directory as an argument.
+It calls `make` twice, first with `ll=1` then a second time w/ `llg=1
+`. It generates two files that contain the list of `.ll` and `.llg` files
+The `.ll` files are the basic block (BB) nodes each containting the code for each node
+the BB's are named with numeric labels, and each line of the node is a unique register name (e.g.):
 
+```
 9:
   %10 = sext i32 %3 to i64
   %11 = getelementptr inbounds i8*, i8** %2, i64 %10
   %12 = sext i32 %1 to i64
   %13 = getelementptr inbounds i8*, i8** %0, i64 %12
   br label %14
+```
 
-The .llg files are similar except that they have debugging information that points each BB to a set of lines in the original source code.
+The `.llg` files are similar except that they have debugging information that points each BB to a set of lines in the original source code.
 
-d2 calls the script llg2cdb on each .llg file to create a .csv file used as a mapping from BB to the C source line at the end
+`d2` calls the script `llg2cdb` on each `.llg` file to create a `.csv` file used as a mapping from BB to the C source line later in the process.
 
-3) for each .ll file, we call an llvm utility (opt) to calculate the number (nesting) of each BB in loops
-opt -analyze --loops
+3) For each `.ll` file, we call an llvm utility (`opt`) to calculate the number (nesting) of each BB in loops
+`opt -analyze --loops`
 
-4) each .ll file is parsed into a .dot file via a script called ll2dot
-this custom parser retains information from the .ll file, namely, the function name and label number
-additionally, it keeps track of any functions a BB calls
-and finally what other nodes (BB) the curent BB can branch to
-this forms the directed graph or .dot file which can be directly viewed with the dot utility
+4) Each `.ll` file is parsed into a `.dot` file via a script called `ll2dot`.
+This custom parser retains information from the `.ll` file, namely, the function name and label number.
+Additionally, it keeps track of any functions a BB calls,
[…]
