### 2024-03-03 7ca437d
add additional dependencies and a check for them
