+++ readme.md
@@ -8,6 +8,7 @@ D2 is intended run on a Unix/Linux system. To build D2, cd to the src directory 
 # Dependencies
 D2 requires an LLVM compiler (e.g. `clang`) and `opt`. (I have only tested with clang)
 It is not required to switch your application to use clang, it is only used for SB intentification.
+Other dependencies: awk bash cpio find grep python sed sort
 
 # Example usage
 
