### 2024-03-03 dtkerns/d2#5 @dtkerns
This error concerns me:
```
/home/root/d2/bin/d2: line 14: [: 14.0.0: integer expression expected
```
That indicates an issue reading the .config file created when you run `make` in the `d2/src`
I made a small change to the d2 script, to be more explicit about where to run `make`
Can you verify (show me here) the contents of your `.config` file
Also, it looks like I make a bad assumption about the availability of `cpio`. I'll list and check for additional dependencies.
