### 2024-03-03 dtkerns/d2#5 @manuel-g-castro
Sure! Here is my `./bin/.config`:

```
root@6ce3f6fc9c55:/home/root/d2/bin# cat .config
LLVMVER=14.0.0
NEWPASSMANAGER=implemented
```
