### 2024-03-03 dtkerns/d2#5 @dtkerns
I just pushed some fixes, should print a warning about LLVM version < 19. please pull, and re-try. thanks
