### 2024-03-03 dtkerns/d2#5 @manuel-g-castro
Hey, David! Thanks for the changes.

I have recreated the container with the ubuntu 22.04 image, and noticed one thing that I forgot to mention: the `opt` dependency installs at `/usr/bin/opt-14`, instead of `/usr/bin/opt`, as expected by your Makefile. I had to create a symlink `/usr/bin/opt` in order to make to work. **I would suggest checking your package manager where and how is it installed and commenting it on the `readme.md`**.

With that done, as the image comes without a compiler, I only installed `clang` for the testing. I found out that the compilation of `d2` expects `g++`, which I tried to amend by executing `make CXX=/usr/bin/clang`. Unfortunately, it failed in the linking stage. Then, I am afraid that `d2`'s compilation requires GNU's compiler, **which should be in the dependency list**.

Also, which version of `python` I need to install? I made a symlink of 3.10.12 to `/usr/bin/python`, and it worked. **But I believe it should be clear at least if its 2.7 or 3+**.

With the current changes, I am on commit `4c071`, and the full installed software dependency, the example worked without problems. Although I find odd that it is complaining that I have an old version of LLVM, 14, and that I need to include `-O1` flag in both app's `CFLAGS`, which I have done. I do not think this is due to your fault. Maybe `clang`'s aptitude package is still shipping with an old, buggy, version.

Finally, with all the bold comments addressed, for my side we can close this issue.

For reference, I am including the files created by `d2` (after removing functions without I/O calls, as pointed out both in the paper and in the `step-by-step.md`).

```
root@aa99ffb7b85f:/home/d2/output# find . -type f -exec ls -l {} \;
-rw-r--r-- 1 root root 90 Mar  3 17:02 ./app1/app1.bb/entry_rotf
-rw-r--r-- 1 root root 173 Mar  3 17:02 ./app1/app1.bb/L_rotf_16.norm
-rw-r--r-- 1 root root 173 Mar  3 17:02 ./app1/app1.bb/L_main_20.norm
-rw-r--r-- 1 root root 19 Mar  3 17:02 ./app1/app1.bb/entry_rotf.norm
-rw-r--r-- 1 root root 439 Mar  3 17:02 ./app1/app1.bb/L_rotf_21
-rw-r--r-- 1 root root 107 Mar  3 17:02 ./app1/app1.bb/L_rotf_14
-rw-r--r-- 1 root root 18 Mar  3 17:02 ./app1/app1.bb/L_rotf_3.norm
-rw-r--r-- 1 root root 46 Mar  3 17:02 ./app1/app1.bb/L_rotf_14.norm
-rw-r--r-- 1 root root 175 Mar  3 17:02 ./app1/app1.bb/L_main_6.norm
-rw-r--r-- 1 root root 172 Mar  3 17:02 ./app1/app1.bb/L_main_3
-rw-r--r-- 1 root root 193 Mar  3 17:02 ./app1/app1.bb/L_main_40.norm
-rw-r--r-- 1 root root 88 Mar  3 17:02 ./app1/app1.bb/L_main_37.norm
-rw-r--r-- 1 root root 46 Mar  3 17:02 ./app1/app1.bb/L_main_18.norm
-rw-r--r-- 1 root root 353 Mar  3 17:02 ./app1/app1.bb/L_rotf_21.norm
-rw-r--r-- 1 root root 113 Mar  3 17:02 ./app1/app1.bb/L_main_42.norm
-rw-r--r-- 1 root root 439 Mar  3 17:02 ./app1/app1.bb/L_main_8
-rw-r--r-- 1 root root 254 Mar  3 17:02 ./app1/app1.bb/L_main_40
-rw-r--r-- 1 root root 79 Mar  3 17:02 ./app1/app1.bb/L_rotf_3
-rw-r--r-- 1 root root 179 Mar  3 17:02 ./app1/app1.bb/L_main_42
-rw-r--r-- 1 root root 19 Mar  3 17:02 ./app1/app1.bb/entry_main.norm
-rw-r--r-- 1 root root 431 Mar  3 17:02 ./app1/app1.bb/L_rotf_4
-rw-r--r-- 1 root root 367 Mar  3 17:02 ./app1/app1.bb/L_main_8.norm
-rw-r--r-- 1 root root 118 Mar  3 17:02 ./app1/app1.bb/entry_main
-rw-r--r-- 1 root root 236 Mar  3 17:02 ./app1/app1.bb/L_main_6
-rw-r--r-- 1 root root 441 Mar  3 17:02 ./app1/app1.bb/L_main_25
-rw-r--r-- 1 root root 114 Mar  3 17:02 ./app1/app1.bb/L_main_3.norm
-rw-r--r-- 1 root root 354 Mar  3 17:02 ./app1/app1.bb/L_main_25.norm
-rw-r--r-- 1 root root 152 Mar  3 17:02 ./app1/app1.bb/L_main_37
-rw-r--r-- 1 root root 242 Mar  3 17:02 ./app1/app1.bb/L_main_20
-rw-r--r-- 1 root root 370 Mar  3 17:02 ./app1/app1.bb/L_rotf_4.norm
-rw-r--r-- 1 root root 239 Mar  3 17:02 ./app1/app1.bb/L_rotf_16
-rw-r--r-- 1 root root 107 Mar  3 17:02 ./app1/app1.bb/L_main_18
-rw-r--r-- 1 root root 1665 Mar  3 17:02 ./app1/app1.sb
-rw-r--r-- 1 root root 18152 Mar  3 17:01 ./app1/app1.llg
-rw-r--r-- 1 root root 1110 Mar  3 17:01 ./app1/app1.dot
-rw-r--r-- 1 root root 260 Mar  3 17:02 ./app1/app1.db.d/g0007.dot
-rw-r--r-- 1 root root 225 Mar  3 17:02 ./app1/app1.db.d/g0004.dot
-rw-r--r-- 1 root root 179 Mar  3 17:02 ./app1/app1.db.d/g0005.dot
-rw-r--r-- 1 root root 235 Mar  3 17:02 ./app1/app1.db.d/g0003.dot
-rw-r--r-- 1 root root 222 Mar  3 17:02 ./app1/app1.db.d/g0001.dot
-rw-r--r-- 1 root root 181 Mar  3 17:02 ./app1/app1.db.d/g0006.dot
-rw-r--r-- 1 root root 207 Mar  3 17:02 ./app1/app1.db.d/g0000.dot
-rw-r--r-- 1 root root 179 Mar  3 17:02 ./app1/app1.db.d/g0002.dot
-rw-r--r-- 1 root root 5841 Mar  3 17:01 ./app1/app1.ll
-rw-r--r-- 1 root root 0 Mar  3 17:01 ./app1/app1.loop
-rw-r--r-- 1 root root 4 Mar  3 17:02 ./app1/app1.const
-rw-r--r-- 1 root root 34 Mar  3 17:02 ./bbselected
-rw-r--r-- 1 root root 68 Mar  3 17:02 ./bbsel_norm
-rw-r--r-- 1 root root 1050 Mar  3 17:02 ./lobb
-rw-r--r-- 1 root root 3359 Mar  3 17:02 ./bbdb
-rw-r--r-- 1 root root 25 Mar  3 17:02 ./bbsel_norm.bb
-rw-r--r-- 1 root root 0 Mar  3 17:02 ./loopdb
-rw-r--r-- 1 root root 30 Mar  3 17:01 ./loll
-rw-r--r-- 1 root root 460 Mar  3 17:02 ./bbsbdb
-rw-r--r-- 1 root root 183 Mar  3 17:02 ./app2/app2.db.d/g0003.dot
-rw-r--r-- 1 root root 225 Mar  3 17:02 ./app2/app2.db.d/g0001.dot
-rw-r--r-- 1 root root 222 Mar  3 17:02 ./app2/app2.db.d/g0000.dot
-rw-r--r-- 1 root root 179 Mar  3 17:02 ./app2/app2.db.d/g0002.dot
-rw-r--r-- 1 root root 159 Mar  3 17:02 ./app2/app2.bb/L_main_16
-rw-r--r-- 1 root root 331 Mar  3 17:02 ./app2/app2.bb/L_rotf_6
-rw-r--r-- 1 root root 167 Mar  3 17:02 ./app2/app2.bb/entry_rotf
-rw-r--r-- 1 root root 241 Mar  3 17:02 ./app2/app2.bb/L_rotf_16.norm
-rw-r--r-- 1 root root 177 Mar  3 17:02 ./app2/app2.bb/L_main_5.norm
-rw-r--r-- 1 root root 68 Mar  3 17:02 ./app2/app2.bb/entry_rotf.norm
-rw-r--r-- 1 root root 200 Mar  3 17:02 ./app2/app2.bb/L_main_62
-rw-r--r-- 1 root root 227 Mar  3 17:02 ./app2/app2.bb/L_main_35.norm
-rw-r--r-- 1 root root 127 Mar  3 17:02 ./app2/app2.bb/L_rotf2_5
-rw-r--r-- 1 root root 385 Mar  3 17:02 ./app2/app2.bb/L_rotf_23
-rw-r--r-- 1 root root 280 Mar  3 17:02 ./app2/app2.bb/L_main_44
-rw-r--r-- 1 root root 92 Mar  3 17:02 ./app2/app2.bb/L_rotf_13.norm
-rw-r--r-- 1 root root 68 Mar  3 17:02 ./app2/app2.bb/entry_rotf2.norm
-rw-r--r-- 1 root root 237 Mar  3 17:02 ./app2/app2.bb/L_main_5
-rw-r--r-- 1 root root 443 Mar  3 17:02 ./app2/app2.bb/L_main_50
-rw-r--r-- 1 root root 194 Mar  3 17:02 ./app2/app2.bb/L_main_66.norm
-rw-r--r-- 1 root root 154 Mar  3 17:02 ./app2/app2.bb/L_rotf2_13
-rw-r--r-- 1 root root 254 Mar  3 17:02 ./app2/app2.bb/L_main_66
-rw-r--r-- 1 root root 95 Mar  3 17:02 ./app2/app2.bb/L_main_41.norm
-rw-r--r-- 1 root root 236 Mar  3 17:02 ./app2/app2.bb/L_main_3
-rw-r--r-- 1 root root 95 Mar  3 17:02 ./app2/app2.bb/L_main_16.norm
-rw-r--r-- 1 root root 276 Mar  3 17:02 ./app2/app2.bb/L_rotf_6.norm
-rw-r--r-- 1 root root 279 Mar  3 17:02 ./app2/app2.bb/L_main_19
-rw-r--r-- 1 root root 67 Mar  3 17:02 ./app2/app2.bb/L_rotf2_5.norm
-rw-r--r-- 1 root root 292 Mar  3 17:02 ./app2/app2.bb/L_main_35
-rw-r--r-- 1 root root 92 Mar  3 17:02 ./app2/app2.bb/L_rotf2_13.norm
-rw-r--r-- 1 root root 206 Mar  3 17:02 ./app2/app2.bb/L_main_44.norm
-rw-r--r-- 1 root root 292 Mar  3 17:02 ./app2/app2.bb/L_main_10
-rw-r--r-- 1 root root 331 Mar  3 17:02 ./app2/app2.bb/L_rotf2_6
-rw-r--r-- 1 root root 130 Mar  3 17:02 ./app2/app2.bb/L_main_62.norm
-rw-r--r-- 1 root root 19 Mar  3 17:02 ./app2/app2.bb/entry_main.norm
-rw-r--r-- 1 root root 180 Mar  3 17:02 ./app2/app2.bb/L_main_68
-rw-r--r-- 1 root root 118 Mar  3 17:02 ./app2/app2.bb/entry_main
-rw-r--r-- 1 root root 168 Mar  3 17:02 ./app2/app2.bb/entry_rotf2
-rw-r--r-- 1 root root 309 Mar  3 17:02 ./app2/app2.bb/L_rotf2_16.norm
-rw-r--r-- 1 root root 386 Mar  3 17:02 ./app2/app2.bb/L_main_25
-rw-r--r-- 1 root root 175 Mar  3 17:02 ./app2/app2.bb/L_main_3.norm
-rw-r--r-- 1 root root 298 Mar  3 17:02 ./app2/app2.bb/L_main_25.norm
-rw-r--r-- 1 root root 276 Mar  3 17:02 ./app2/app2.bb/L_rotf2_6.norm
-rw-r--r-- 1 root root 127 Mar  3 17:02 ./app2/app2.bb/L_rotf_5
-rw-r--r-- 1 root root 159 Mar  3 17:02 ./app2/app2.bb/L_main_41
-rw-r--r-- 1 root root 154 Mar  3 17:02 ./app2/app2.bb/L_rotf_13
-rw-r--r-- 1 root root 356 Mar  3 17:02 ./app2/app2.bb/L_main_50.norm
-rw-r--r-- 1 root root 205 Mar  3 17:02 ./app2/app2.bb/L_main_19.norm
-rw-r--r-- 1 root root 227 Mar  3 17:02 ./app2/app2.bb/L_main_10.norm
-rw-r--r-- 1 root root 442 Mar  3 17:02 ./app2/app2.bb/L_rotf2_25
-rw-r--r-- 1 root root 309 Mar  3 17:02 ./app2/app2.bb/L_rotf_16
-rw-r--r-- 1 root root 67 Mar  3 17:02 ./app2/app2.bb/L_rotf_5.norm
-rw-r--r-- 1 root root 381 Mar  3 17:02 ./app2/app2.bb/L_rotf2_16
-rw-r--r-- 1 root root 297 Mar  3 17:02 ./app2/app2.bb/L_rotf_23.norm
-rw-r--r-- 1 root root 355 Mar  3 17:02 ./app2/app2.bb/L_rotf2_25.norm
-rw-r--r-- 1 root root 114 Mar  3 17:02 ./app2/app2.bb/L_main_68.norm
-rw-r--r-- 1 root root 8697 Mar  3 17:01 ./app2/app2.ll
-rw-r--r-- 1 root root 0 Mar  3 17:01 ./app2/app2.loop
-rw-r--r-- 1 root root 1928 Mar  3 17:01 ./app2/app2.dot
-rw-r--r-- 1 root root 809 Mar  3 17:02 ./app2/app2.sb
-rw-r--r-- 1 root root 28860 Mar  3 17:01 ./app2/app2.llg
-rw-r--r-- 1 root root 14 Mar  3 17:02 ./app2/app2.const
-rw-r--r-- 1 root root 1613 Mar  3 17:01 ./bbcdb
-rw-r--r-- 1 root root 32 Mar  3 17:01 ./lollg
-rw-r--r-- 1 root root 27 Mar  3 17:02 ./sbselected
-rw-r--r-- 1 root root 1112 Mar  3 17:02 ./rankorder
-rw-r--r-- 1 root root 27 Mar  3 17:02 ./sbmarkup
-rw-r--r-- 1 root root 60 Mar  3 17:02 ./ronormfiles
```
