### 2024-03-03 90786bc
more specific dependencie checks, more warnings/checks/messages about dependencies
