+++ readme.md
@@ -7,8 +7,10 @@ D2 is intended run on a Unix/Linux system. To build D2, cd to the src directory 
 
 # Dependencies
 D2 requires an LLVM compiler (e.g. `clang`) and `opt`. (I have only tested with clang)
+Some LLVM installations install `opt` as `opt-VV` where VV is a version number. Please resolve with a symbolic link or modify the `d2` script as appropriate.
 It is not required to switch your application to use clang, it is only used for SB intentification.
-Other dependencies: awk bash cpio find grep python sed sort
+Other dependencies:
+`awk bash cpio cut find grep python3 sed sort` and a c++ comipler, either `g++` or `clang++`
 
 # Example usage
 
