### 2024-03-03 dtkerns/d2#5 @dtkerns
hello, thanks for your continued review.

LLVM's installation of `opt` is going to be really tricky and distro dependent. (I did most of the original development of this package under Cent OS7, and switching to RHEL 9 caused a plethora of LLVM issues.) I made a note of the `opt-VV` issue in the readme.md, but without a suggested solution, I may just rely a a human to to correct it as needed. On RHEL 9, LLVM is on version 16 (Cent OS7 was LLVM version 3!) ... unfortunately, I don't expect either to ever update to v19, where LLVM developers are currently at.

d2 is c++ code, so it either needs to be `g++`,  `clang++` or another c++ compiler of your choosing (and definition of `CXX` in the `src/Makefile` . There is nothing in sb.cc that a vanilla c++ compiler can't handle; a C compiler, not so much. ;) I added additional checks in the src/Makefile for this too.

Finally, I specified explicitly `python3` as the required dependency; sorry, I should have known better.

If you're ok with these 3 items as I addressed them, I'll close the issue.
