### 2024-06-09 dtkerns/d2#6 @abhishektiwari
Specify LLVM versions in installation and usage instruction
Got following warning during installation, please specify supported LLVM versions in installation and usage instruction to ensure users can follow the instructions when using LLVM versions < 19 or 19>=.

```
warning: due to a bug in LLVM versions < 19, you will need to add a -O1 (minimum) to CFLAGS in your applications makefiles
```

[For JOSS Review](https://github.com/openjournals/joss-reviews/issues/6825)
