### 2024-06-09 dtkerns/d2#7 @abhishektiwari
Any analysis results to be reported for the paper?
Are there any analysis results reported by the paper as stated in summary section _Our results show that the SB should be taken seriously as a candidate for creating DSAs._. I can't find any benchmark or analysis to support the statement in paper or repository.

[For JOSS Review](https://github.com/openjournals/joss-reviews/issues/6825)
