### 2024-06-09 dtkerns/d2#6 @dtkerns
That's correct, please see issue #3. All versions of LLVM are supported. However, the original work was done under CentOS7 which was/is running LLVM version 3.4.2. (which still runs without issue) While "modernizing" to RHEL9 (which runs LLVM version 16.0.6) the loop bug was discovered and documented. Due to the nature/philosophy of RHEL, it is not likely RHEL 9 will ever go beyond the the 16 series. The LLVM bug was fixed in version 19. Thus, users running LLVM versions > 13 and < 19, will need to add the -O1 (minimum optimization level) in order to successfully get the loop output from the LLVM `opt` command.
