### 2024-06-09 dtkerns/d2#7 @dtkerns
The original version of this paper (paper/full_paper.md) had many of the details you are asking for. However, the initial (pre)reviewers asked me to trim the paper down significantly for submission to JOSS.
