### 2024-06-17 dtkerns/d2#7 @abhishektiwari
Thanks. Much appreciated. Can analysis results be uploaded to Zenodo and referenced in paper? We can ask editors.
