### 2024-06-17 dtkerns/d2#7 @dtkerns
I'm not familiar with Zenodo, but it sounds like a reasonable request.
