+++ paper/appendix.md
@@ -1,0 +1,135 @@
+---
+title: 'Appendix A: Case studies'
+tags:
+ - Domain-specific accelerators
+ - computer-aided design
+ - resource-constrained systems
+ - energy efficient
+authors:
+ - name: David T Kerns
+   affiliation: 1
+ - name: Tosiron Adegbija
+   affiliation: 1
+affiliations:
+- name: Department of Electrical & Computer Engineering, University of Arizona, Tucson, Arizona, USA
+  index: 1
+date: 01 December 2023
+bibliography: paper.bib
+---
+
+# Case Studies
+To demonstrate the use of **D2** and its ease of setup, we used the MiBench benchmark suite [@guthaus01_mibench] since it is popular and features several workloads that are popular in the embedded systems computing space. The MiBench suite is split into six main groups, of the six, we considered `automotive` and `consumer`. Each of those groups has four or five applications, or workloads, ranging from bit manipulation to edge identification in an image. We ran **D2** on the full suite of workloads but because of the wide spectrum and diversity of algorithms implemented, there was minimal opportunity for DSA to be shared amongst the different workloads. There were, however, multiple instances of an SB being shared in multiple places within a given workload that **D2** was able to identify because of the normalization step. Using **D2**, we were able to identify multiple DSAs, six of which we will highlight here since they were ranked high by the **D2** tool.   
+
+Although the tool successfully identified DSAs that significantly sped up the computations in the workloads, the communication between the DSA and the host remained a bottleneck in our case studies. Depending on the interface that is used between the host and the DSA---CPU and FPGA, in our experiments---the overhead of communicating between the host and DSA may be too large to overcome for small accelerators. The most straightforward implementation is to memory map the FPGA\'s input, output, and control registers into the global memory space of the CPU. This option also has the highest overhead to overcome. **D2** has the promise of mitigating this overhead using the superblock granularity for generating DSAs. In our analysis for this paper, we modeled the simple memory-mapped register interface since optimizing the communication interface is outside the scope of this work. While this approach suffices for demonstrating **D2**\'s utility; we note that even better results can be achieved using a higher-performing DMA (Direct Memory Access) or other vendor-specific communication interface between the host and DSA.
+For each of the DSAs below, we used the gem5 [@binkert11_gem5] simulator to calculate the cycle counts of both the original un-accelerated code, as well as the accelerated code with the simulated CPU/FPGA interface. The cycle counts of the actual DSA are from the Vivado simulation with a 100MHz clock; and the host ARM CPU runs at 1GHz. We included the overhead of calling the accelerator from the host, demonstrating the bottleneck of the CPU/DSA interface.
+
+## DSA 1
+The first DSA we will look at was normalized from 4 different SBs. Via the normalization process, as detailed above, we were able to create a DSA that accepted arguments (data) that were passed into the DSA some of which were constants in the original code; in this instance, c1 and c2.
+```c
+int acc1(int c1, int Q00, int c2, int Qs, int Al)
+{
+  int pred;
+  int num = c1 * Q00 * c2;
+  if (num >= 0) {
+    pred = (int) (((Qs<<7)+num) / (Qs<<8));
+    if (Al>0 && pred>=(1<<Al))
+      pred = (1<<Al)-1;
+  } else {
+    pred = (int) (((Qs<<7)-num) / (Qs<<8));
+    if (Al>0 && pred>=(1<<Al))
+      pred = (1<<Al)-1;
+    pred = -pred;
+  }
[…]
