### 2024-06-23 dtkerns/d2#7 @dtkerns
Sorry for the delay. After consulting with my advisor, we came to the the conclusion that an appendix right here would be more appropriate. I've added the file `paper/appendix.md` to address this issue.
