### 2024-07-08 dtkerns/d2#6 @dtkerns
Should I close this? I've been waiting for a comment/confirmation from the originator.
