### 2024-07-08 #6825 @dtkerns
Should I close those two issues? I was waiting for a follow-up comment from the originator.
