### 2024-08-06 dtkerns/d2#8 @abhishektiwari
[JOSS Review] Automated examples to run on every commit as E2E test
Given there are no automated test for this package, I strongly recommend authors adding a Github action which will run the `app1` and `app2` in `example` folder on eery commit.

See [JOSS testing](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) requirements.

> Good: An automated test suite hooked up to continuous integration (GitHub Actions, Circle CI, or similar)
OK: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behavior)
Bad (not acceptable): No way for you, the reviewer, to objectively assess whether the software works

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/6825)
