### 2024-08-06 #6825 @abhishektiwari
@dtkerns I am down to final review criteria - **Automated tests**. Please see my recommendations [here](https://github.com/dtkerns/d2/issues/8). I think this is a small change to automate existing example to run with every commit without too much of rework.
