### 2024-08-06 #6825 @abhishektiwari
@gkthiruvathukal tool has performance benchmark under the `paper/appendix.md` of repository. Is JOSS guidance to register benchmark results with a separate ZENODO archive or its ok to be keep it part of repository ZENODO archive?
