### 2024-08-06 dtkerns/d2#8 @dtkerns
There is a "human in the loop" portion of the script, but I think I can automate it with a few assumptions. I'll get that added ASAP.
