### 2024-08-06 36a0298
add test target to example to support CI/CD
