+++ src/Makefile
@@ -27,4 +27,7 @@ deps:
 	@echo dependency check:
 	@which awk cpio cut clang find grep python3 sed sort ; test $$? = 0 || (echo please install any commands not found; exit 1)
 
-.PHONY: deps versioncheck all
+test: install
+	$(MAKE) -C ../example; 
+
+.PHONY: deps versioncheck all test
