+++ example/Makefile
@@ -8,15 +8,19 @@ export llg
 endif
 
 APPS = app1 app2
+export CFLAGS = -O1
 
 all: $(APPS)
 
 app1 app2:
-	$(MAKE) -C $@ $(MAKECMDGOALS)
+	$(MAKE) -C $@ $(filter-out test,$(MAKECMDGOALS))
 
 clean:
 	for d in $(APPS); do \
 		$(MAKE) -C $$d $@; \
 	done
+
+test: all
+	mkdir /tmp/testout$$; ../bin/d2 -i /tmp/testout$$
 
 .PHONY: app1 app2
