### 2024-08-06 dtkerns/d2#8 @dtkerns
As a first step, I've added a `test` target in the example directory Makefile. That should get me to "OK" ... in the coming days (as time allows) I will attempt to automate it to full blown CI/CD.
