### 2024-08-08 dtkerns/d2#8 @dtkerns
I'm a CI/CD noob, can I get an assist with the yaml?
