+++ .github/workflows/test.yml
@@ -8,4 +8,4 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v4
       - name: Build
-        run: pwd; ls; make -C examples test
+        run: make -C example test
