### 2024-08-08 9bbfca3
build/test from src, don't need to do all rule in example for test
