+++ .github/workflows/test.yml
@@ -8,4 +8,4 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v4
       - name: Build
-        run: make -C example test
+        run: make -C src test
