### 2024-08-23 #6825 @gkthiruvathukal
@abhishektiwari As @danielskatz mentioned, it is ok.  Any service with a stable DOI will do.
