### 2024-08-23 dtkerns/d2#4 @manuel-g-castro
Sorry! After a long hiatus, I am back at your submission.

I noticed a tiny typo at line 13 of the `readme.md` where it says "(...) and a c++ **comipler** (...)".

I suppose this will be swiftly patched, hence I have already checkout the box of the intallation guide.
