### 2024-08-27 dtkerns/d2#9 @manuel-g-castro
Improve single phrase in step-by-step and mention appendix
Hello, @dtkerns:

I really appreciate having a for dummies (like me) step by step for reproducing our software. I found a little issue at the end of line 10 and begining of 11 of `step-by-step.md`. Where you write "build a `.ll` file insted also ...". Maybe you meant to say that `.o` and also `.ll`  files will be build. But I am unsure.

I also like a lot your `appendix.md`, which is really nicely writen and interesting. Maybe it is worth it to mention the existance of this file on the `readme.md`?
