### 2024-08-27 #6825 @manuel-g-castro
hey @dtkerns I will just wait for your comments [on the issue that I just opened](https://github.com/dtkerns/d2/issues/9).

With that done, I think, there is nothing more from my side.
