### 2024-08-27 dtkerns/d2#9 @dtkerns
it actually does not build a `.o` file ... from the `clang` man page:
```
-emit-llvm
              Generate output files in LLVM formats, suitable for link time optimization.  When used  with  -S  this
              generates  LLVM intermediate language assembly files...
the flags `-S -emit-llvm` cause the `clang` compiler to only do the the "first" step of converting the C into `clang`'s intermediate
