### 2024-08-27 dtkerns/d2#9 @dtkerns
the appendix.md was part of the original paper. unfortunately, it caused the paper to exceed the length of the JOSS guidelines. but one of the reviewers specifically asked to see (quantitative) results, so I added it back as an appendix.

I'm open to suggestions on a good way to refer to it in the readme.
