### 2024-08-27 bae0514
Update step-by-step.md
fix run on sentence.
