### 2024-08-27 dtkerns/d2#9 @manuel-g-castro
> it actually does not build a `.o` file ... from the `clang` man page:
>
> ```
> -emit-llvm
>               Generate output files in LLVM formats, suitable for link time optimization.  When used  with  -S  this
>               generates  LLVM intermediate language assembly files...
> ```
>
> so we stop short of generating the `.o` files with those flags.

So there is an extra `also` there on the test. Cool! Solved.

> the appendix.md was part of the original paper. unfortunately, it caused the paper to exceed the length of the JOSS guidelines. but one of the reviewers specifically asked to see (quantitative) results, so I added it back as an appendix.
>
> I'm open to suggestions on a good way to refer to it in the readme.

Hi! Maybe include an example section? Or performance benchmark... I am unsure about what works best, but feels a pity to have it there but you have to dig the repository to find it.
