### 2024-08-27 dtkerns/d2#9 @dtkerns
how's this wording?

# Benchmarks

To meet JOSS length guidelines, the quantitative analysis was removed from the paper, but is available as an appendix. see paper/appendix.md
