### 2024-08-27 9d11a87
Update readme.md
include pointer to quantitative analysis
