### 2024-08-27 dtkerns/d2#9 @dtkerns
Quantitative analysis is available as an appendix. see paper/appendix.md
