### 2024-08-27 05450b1
Update step-by-step.md
fix wording in step one, to clearly state the purpose of the flags passed to clang
