### 2024-08-27 dtkerns/d2#9 @tosiron
No need to include "to meet Joss length guidelines".

Tosi
