### 2024-08-27 #6825 @dtkerns
I'll let you close the issue when you're happy with the updates (I'm a
believer in "the issue opener must be satisfied to close")

On Tue, Aug 27, 2024 at 5:57 AM Manuel G. Marciani ***@***.***>
wrote:

> hey @dtkerns <https://github.com/dtkerns> I will just wait for your
> comments on the issue that I just opened
> <https://github.com/dtkerns/d2/issues/9>.
>
> With that done, I think, there is nothing more from my side.
>
