### 2024-08-27 #6825 @manuel-g-castro
> I'll let you close the issue when you're happy with the updates (I'm a believer in "the issue opener must be satisfied to close")
> […](#)
> On Tue, Aug 27, 2024 at 5:57 AM Manuel G. Marciani ***@***.***> wrote: hey @dtkerns <https://github.com/dtkerns> I will just wait for your comments on the issue that I just opened <[dtkerns/d2#9](https://github.com/dtkerns/d2/issues/9)>. With that done, I think, there is nothing more from my side. — Reply to this email directly, view it on GitHub <[#6825 (comment)](https://github.com/openjournals/joss-reviews/issues/6825#issuecomment-2312491360)>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/ADVJRKDYFE5LEOWLVD4E2R3ZTRZTXAVCNFSM6AAAAABIVCPMRCVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDGMJSGQ4TCMZWGA> . You are receiving this because you were mentioned.Message ID: ***@***.***>

Done! I have nothing else to comment. Nice work!
