### 2024-10-11 #6825 @danielskatz
@dtkerns - I'll take over at this point.  I'll next proofread the paper (either Sunday or Monday, most likely), and let you know if I see anything to be changed, then give you instructions on packaging up the files and moving this onward.
