### 2024-10-12 #6825 @gkthiruvathukal
@danielskatz Thanks! Per https://github.com/openjournals/joss-reviews/issues/6825#issuecomment-2362679375, I was ready to move forward. My mention of the checklist was a momentary brain lapse where I still was thinking as a track EIC! So please proceed.
