### 2024-10-13 #6825 @danielskatz
@gkthiruvathukal - no, you were thinking correctly, in that a checklist would have made sense for you to request to get it ready for me, but since I wasn't sure when you might get to this, I just decided to go ahead, and do things in a different order since I was going to take over the final editor steps as well as doing the AEiC steps.
