### 2024-10-13 8855a61
minor changes in paper.md
