### 2024-10-13 #6825 @danielskatz
👋 @dtkerns - I've suggested a few minor changes to your paper in https://github.com/dtkerns/d2/pull/10 - please merge this, or let me know what you disagree with, then we can continue.
