### 2024-10-13 #6825 @dtkerns
changes merged (thanks) also updated to ***@***.*** per
deprecation build error mesage

On Sun, Oct 13, 2024 at 5:01 AM Daniel S. Katz ***@***.***>
wrote:

> 👋 @dtkerns <https://github.com/dtkerns> - I've suggested a few minor
> changes to your paper in dtkerns/d2#10
> <https://github.com/dtkerns/d2/pull/10> - please merge this, or let me
> know what you disagree with, then we can continue.
>
