### 2024-10-13 #6825 @danielskatz
You can do this however you would like, but the easiest way is probably to use the GitHub - Zenodo link described in https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content
