### 2024-10-13 #6825 @dtkerns
DOI

10.5281/zenodo.13926674

On Sun, Oct 13, 2024 at 11:18 AM Daniel S. Katz ***@***.***>
wrote:

> You can do this however you would like, but the easiest way is probably to
> use the GitHub - Zenodo link described in
> https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content
>
