### 2024-10-13 #6825 @dtkerns
oh , and zenodo made me make a new tag, so v1.0.0-pub

On Sun, Oct 13, 2024 at 11:35 AM david kerns ***@***.***>
wrote:

> DOI
>
> 10.5281/zenodo.13926674
>
>
> On Sun, Oct 13, 2024 at 11:18 AM Daniel S. Katz ***@***.***>
> wrote:
>
>> You can do this however you would like, but the easiest way is probably
>> to use the GitHub - Zenodo link described in
>> https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content
>>
>> —
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/6825#issuecomment-2409074746>,
>> or unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/ADVJRKH7FV3TAJ7QYL5RTATZ3K2PDAVCNFSM6AAAAABIVCPMRCVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDIMBZGA3TINZUGY>
>> .
>> You are receiving this because you were mentioned.Message ID:
>> ***@***.***>
>>
>
