### 2024-10-13 #6825 @danielskatz
@dtkerns - please change the metadata for the archive so that the title and authors match those of the paper. (You don't need to create a new version or new DOI for this; you can do it via editing the Zenodo archive's metadata - see https://help.zenodo.org/docs/deposit/manage-records/)
