### 2024-10-13 #6825 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
