### 2024-10-13 #6825 @dtkerns
metadata updated

On Sun, Oct 13, 2024 at 11:47 AM The Open Journals editorial robot <
***@***.***> wrote:

> 👋 @openjournals/csism-eics
> <https://github.com/orgs/openjournals/teams/csism-eics>, this paper is
> ready to be accepted and published.
>
> Check final proof 👉📄 Download article
> <https://raw.githubusercontent.com/openjournals/joss-papers/joss.06825/joss.06825/10.21105.joss.06825.pdf>
>
> If the paper PDF and the deposit XML files look good in
> openjournals/joss-papers#5994
> <https://github.com/openjournals/joss-papers/pull/5994>, then you can now
> move forward with accepting the submission by compiling again with the
> command @editorialbot accept
>
