### 2024-10-13 #6825 @danielskatz
I see the authors as partially updated, but not the title. And the authors aren't using names (and ORCIDs if desired), but still github usernames
