### 2024-10-13 #6825 @dtkerns
sorry I missed those. done
I noticed that the Zenodo entry defaulted to a Creative Commons 4.0 but On
github I specified an MIT License ... are those different things, or do
they need to be consistent?

On Sun, Oct 13, 2024 at 12:44 PM Daniel S. Katz ***@***.***>
wrote:

> I see the authors as partially updated, but not the title. And the authors
> aren't using names (and ORCIDs if desired), but still github usernames
> Screenshot.2024-10-13.at.14.43.33.png (view on web)
> <https://github.com/user-attachments/assets/ecfe24ae-d168-4403-b519-fee9309d9afe>
>
