### 2024-10-13 #6825 @danielskatz
Congratulations to @dtkerns (David Kerns) and co-author on your publication!!
