### 2024-10-13 #6825 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
