### 2025-05-16 #6825 @ritika-giri
Hi @dtkerns - you can submit in the talk, poster or any other track except the paper track. Let me know if you have any other questions or concerns or would like help with the submission guidelines.
